frontend_service/requirements.txt
streamlit==1.43.2
requests==2.32.3
pillow==10.2.0

